+++ MethodsTests/src/test/java/TESTS/FireFox_Browser_Test.java
@@ -1,0 +1,24 @@
+package TESTS;
+
+import org.junit.Test;
+
+import globalized.GlobalizedAllPage;
+
+public class FireFox_Browser_Test extends GlobalizedAllPage { 
+	
+	String URL = "https://www.google.com";
+	
+	@Test
+	public void Test1()
+		{
+			
+		}
+	
+	@Test
+	public void Test2()
+		{
+	
+		}
+	
+	
+}
