+++ MethodsTests/src/test/java/globalized/GlobalizedAllPage.java
@@ -1,0 +1,11 @@
+package globalized;
+
+import Healper.DriverUtility;
+import Moduals.WaitUtility;
+import Moduals.WindowUtility;
+
+public class GlobalizedAllPage extends DriverUtility {
+	
+		public static WaitUtility UtilWait = new WaitUtility();
+		public static WindowUtility UtilWin = new WindowUtility();
+}
